chore: move formatting logic to foundry-macros
