+++ common/src/lib.rs
@@ -14,11 +14,12 @@ pub mod fs;
 pub mod provider;
 pub mod selectors;
 pub mod shell;
-pub use provider::*;
 pub mod term;
 pub mod traits;
+pub mod transactions;
+
 pub use constants::*;
 pub use contracts::*;
+pub use provider::*;
 pub use traits::*;
-pub mod transactions;
 pub use transactions::*;
