+++ macros/src/fmt/mod.rs
@@ -5,3 +5,6 @@ pub use ui::*;
 
 mod token;
 pub use token::*;
+
+mod console_fmt;
+pub use console_fmt::{console_format, ConsoleFmt, FormatSpec};
