+++ macros/src/fmt/ui.rs
@@ -1,20 +1,18 @@
-//! Contains a helper pretty() function to print human redeable string versions of usual ethers
-//! types
+//! Helper trait and functions to format ethers types.
+
 use ethers_core::{
     types::*,
     utils::{hex, to_checksum},
 };
 use serde::Deserialize;
 
-use crate::TransactionReceiptWithRevertReason;
-
 /// length of the name column for pretty formatting `{:>20}{value}`
 const NAME_COLUMN_LEN: usize = 20usize;
 
+/// Helper trait to format ethers types.
 ///
-/// Uifmt is a helper trait to format the usual ethers types
-/// It offers a `pretty()` function that returns a human readable String of the value
-/// # Example
+/// # Examples
+///
 /// ```
 /// use foundry_common::fmt::UIfmt;
 /// let boolean: bool = true;
@@ -102,10 +100,17 @@ impl<T: UIfmt> UIfmt for Option<T> {
 impl<T: UIfmt> UIfmt for Vec<T> {
     fn pretty(&self) -> String {
         if !self.is_empty() {
-            format!(
-                "[\n{}]",
-                self.iter().fold("".to_string(), |s, x| s + tab_paragraph(x.pretty()).as_str())
-            )
+            let mut s = String::with_capacity(self.len() * 64);
+            s.push_str("[\n");
+            for item in self {
+                for line in item.pretty().lines() {
+                    s.push('\t');
+                    s.push_str(line);
+                    s.push('\n');
+                }
+            }
+            s.push(']');
+            s
         } else {
             "[]".to_string()
         }
@@ -143,21 +148,6 @@ type                    {}",
             self.transaction_index.pretty(),
             self.transaction_type.pretty()
         )
-    }
-}
-
-impl UIfmt for TransactionReceiptWithRevertReason {
-    fn pretty(&self) -> String {
-        if let Some(ref revert_reason) = self.revert_reason {
-            format!(
-                "{}
-revertReason            {}",
-                self.receipt.pretty(),
-                revert_reason
-            )
-        } else {
-            self.receipt.pretty()
-        }
     }
 }
 
@@ -341,10 +331,6 @@ value                {}{}",
     }
 }
 
-fn tab_paragraph(paragraph: String) -> String {
-    paragraph.lines().fold("".to_string(), |acc, x| acc + "\t" + x + "\n")
-}
-
 /// Convert a U256 to bytes
 pub fn to_bytes(uint: U256) -> Bytes {
     let mut buffer: [u8; 4 * 8] = [0; 4 * 8];
@@ -407,36 +393,6 @@ pub fn get_pretty_block_attr<TX>(block: &Block<TX>, attr: &str) -> Option<String
             }
             None
         }
-    }
-}
-
-/// Returns the ``UiFmt::pretty()` formatted attribute of the transaction receipt
-pub fn get_pretty_tx_receipt_attr(
-    receipt: &TransactionReceiptWithRevertReason,
-    attr: &str,
-) -> Option<String> {
-    match attr {
-        "blockHash" | "block_hash" => Some(receipt.receipt.block_hash.pretty()),
-        "blockNumber" | "block_number" => Some(receipt.receipt.block_number.pretty()),
-        "contractAddress" | "contract_address" => Some(receipt.receipt.contract_address.pretty()),
-        "cumulativeGasUsed" | "cumulative_gas_used" => {
-            Some(receipt.receipt.cumulative_gas_used.pretty())
-        }
-        "effectiveGasPrice" | "effective_gas_price" => {
-            Some(receipt.receipt.effective_gas_price.pretty())
-        }
-        "gasUsed" | "gas_used" => Some(receipt.receipt.gas_used.pretty()),
-        "logs" => Some(receipt.receipt.logs.pretty()),
-        "logsBloom" | "logs_bloom" => Some(receipt.receipt.logs_bloom.pretty()),
-        "root" => Some(receipt.receipt.root.pretty()),
-        "status" => Some(receipt.receipt.status.pretty()),
-        "transactionHash" | "transaction_hash" => Some(receipt.receipt.transaction_hash.pretty()),
-        "transactionIndex" | "transaction_index" => {
-            Some(receipt.receipt.transaction_index.pretty())
-        }
-        "type" | "transaction_type" => Some(receipt.receipt.transaction_type.pretty()),
-        "revertReason" | "revert_reason" => Some(receipt.revert_reason.pretty()),
-        _ => None,
     }
 }
